[DDD light] Updated the card service to go hand in hand with the new architecture

+++ backend/ArhiTodo.Application/DTOs/Board/BoardGetDto.cs
@@ -4,7 +4,7 @@ namespace ArhiTodo.Application.DTOs.Board;
 
 public class BoardGetDto
 {
-    public long BoardId { get; set; }
+    public int BoardId { get; set; }
         
     public required string BoardName { get; set; }
