+++ backend/ArhiTodo.Domain/Entities/Auth/ProjectManager.cs
@@ -4,7 +4,7 @@ namespace ArhiTodo.Domain.Entities.Auth;
 
 public class ProjectManager
 {
-    public long ProjectId { get; private set; }
+    public int ProjectId { get; private set; }
     public Project Project { get; } = null!;
     
     public Guid UserId { get; private set; }
@@ -12,7 +12,7 @@ public class ProjectManager
 
     private ProjectManager() { }
     
-    public ProjectManager(long projectId, Guid userId)
+    public ProjectManager(int projectId, Guid userId)
     {
         ProjectId = projectId;
         UserId = userId;
