+++ backend/ArhiTodo.Domain/Entities/Kanban/Checklist.cs
@@ -2,9 +2,9 @@
 
 public class Checklist
 {
-    public long CardId { get; private set; }
+    public int CardId { get; private set; }
     
-    public long ChecklistId { get; private set; }
+    public int ChecklistId { get; private set; }
     public string ChecklistName { get; private set; } = string.Empty;
 
     private readonly List<ChecklistItem> _checklistItems = [];
